apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "baboom"
version = "0.1.0"
description = ""
requires-python = ">= 3.14"

dependencies = [
  "gunicorn",
  "whitenoise",
  "psycopg",
  "requests",
  "Pillow",
  "django>=6.0",
  "django-filter",
  "django-simple-history",
  "django-treebeard",
  "django-nested-admin",
  "django-environ",
  "django-htmx",
  "django-widget-tweaks",
  "celery",
  "playwright",
  "beautifulsoup4",
  "lxml",
  "urllib3",
  "curl_cffi",
  "django-celery-results",
  "django-celery-beat",
  "django-components",
  "strawberry-graphql-django",
  "strawberry-graphql",
  "extruct",
  "w3lib",
  "pydantic-ai",
  "logfire",
  "pydantic",
  "python-dotenv",
  "dagster",
  "dagster-webserver",
]

[project.urls]
Repository = "https://github.com/rflemerson/baboom"

[project.optional-dependencies]
dev = [
  "mypy",
  "django-stubs",
  "django-stubs-ext",
  "ruff",
  "pre-commit",
  "djlint",
  "djhtml",
  "deptry",
  "vulture",
  "safety",
  "factory-boy",
  "pytest",
  "pytest-cov",
  "pytest-django",
  "pytest-asyncio",
  "pytest-sugar",
]

[tool.ruff]
line-length = 88
target-version = "py314"
exclude = [
    ".venv",
    ".git",
    "migrations",
    "docs",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "DJ",  # django specific
    "B",   # flake8-bugbear
    "S",   # flake8-bandit (security)
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RET", # flake8-return
    "ERA", # eradicate (commented code)
    "T20", # flake8-print
    "RUF", # ruff-specific rules
    "A",   # flake8-builtins
    "C90", # mccabe complexity
    "PL",  # pylint
    "N",   # pep8-naming
    "D",   # pydocstyle
]
ignore = [
    "E501",    # Line length
    "RUF012",  # Pydantic ClassVar
    "PLR2004", # Magic values
    "ISC001",  # Formatter conflict
    # Docstrings ignores
    "D100", "D104", "D107", "D203", "D212",
]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.pylint]
max-args = 8
max-branches = 15
max-statements = 60

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"baboom/settings/__init__.py" = ["F403", "F405"]

[tool.mypy]
plugins = ["mypy_django_plugin.main", "strawberry.ext.mypy_plugin", "pydantic.mypy"]
check_untyped_defs = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "baboom.settings"

[tool.djlint]
profile = "django"
extension = "html"
ignore = "H030,H031"
indent = 2
blank_line_after_tag = "load,extends,include"

[tool.deptry]
ignore_notebooks = true
extend_exclude = ["infra", "assets", "docs", "services/agents/agents/tests"]
ignore = ["DEP002"]

[tool.vulture]
paths = ["core", "scrapers", "baboom"]
exclude = ["*/migrations/*", ".venv/", "docs/", "*/settings/*"]
ignore_decorators = [
    "@admin.register",
    "@receiver",
    "@app.task",
    "@task",
    "@shared_task",
    "@asset",
    "@op",
    "@job",
    "@sensor",
    "@schedule",
    "@repository"
]
ignore_names = ["Meta", "get_*", "clean_*", "save_*", "form_valid", "form_invalid", "dispatch", "get_queryset", "get_context_data", "setUp", "tearDown", "test_*", "sender", "args", "cls"]
min_confidence = 80

[tool.interrogate]
ignore-init-module = true
ignore-init-method = true
ignore-magic = true
ignore-semiprivate = true
ignore-private = true
ignore-nested-functions = true
fail-under = 80
exclude = ["setup.py", "docs", "build", "tests", "migrations", "temp", "tmp"]
verbose = 2

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "baboom.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "--reuse-db --nomigrations"
asyncio_mode = "auto"
